requirements.txt
Flask==1.1.1
nltk==3.6.2
pandas==1.2.4
python-dotenv==0.17.1
joblib==1.0.1
scikit-learn==0.22.2
spacy==3.0.6
spacy-legacy==3.0.5
click==7.1.2
dill==0.3.3